Refactoring plugin dir structure to fix standard `ckanext/plugin_name` format.
flake 8, dir refactor fixup
