+++ setup.cfg
@@ -1,20 +1,22 @@
-[compile_catalog]
-domain = ckanext-ytp-request
-directory = i18n
-statistics = true
-
 [extract_messages]
+keywords = translate isPlural
 add_comments = TRANSLATORS:
-output_file = i18n/ckanext-ytp-request.pot
+output_file = ckanext/ytp_request/i18n/ckanext-ytp-request.pot
 width = 80
 
 [init_catalog]
 domain = ckanext-ytp-request
-input_file = i18n/ckanext-ytp-request.pot
-output_dir = i18n
+input_file = ckanext/ytp_request/i18n/ckanext-ytp-request.pot
+output_dir = ckanext/ytp_request/i18n
 
 [update_catalog]
 domain = ckanext-ytp-request
-input_file = i18n/ckanext-ytp-request.pot
-output_dir = i18n
-previous = true
+input_file = ckanext/ytp_request/i18n/ckanext-ytp-request.pot
+output_dir = ckanext/ytp_request/i18n
+
+[compile_catalog]
+domain = ckanext-ytp-request
+directory = ckanext/ytp_request/i18n
+statistics = true
+
+[tool:pytest]
